google_cloud_sync_app/requirements.txt
google-cloud-speech>=2.33.0
google-cloud-translate>=3.21.1
yt-dlp>=2026.03.01
